requirements.txt
# =============================================================================
# Agentic Email Security System - Python Dependencies
# =============================================================================

# --- ML & AI Libraries ---
transformers>=4.36.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0
pandas>=2.1.0
numpy>=1.26.0
torch>=2.1.0
scipy>=1.11.0
joblib>=1.3.0

# --- NLP ---
nltk>=3.8.0
spacy>=3.7.0

# --- Security & Email Analysis ---
email-validator>=2.1.0
tldextract>=5.1.0
python-whois>=0.9.0
yara-python>=4.3.0; platform_system!="Windows"
dnspython>=2.4.0
pefile>=2023.2.7
oletools>=0.60.0
ssdeep>=3.4; platform_system!="Windows"

# --- API Framework ---
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pika>=1.3.2
redis>=5.0.1
psycopg2-binary>=2.9.9
beautifulsoup4>=4.12.3
python-Levenshtein>=0.25.1
docker>=7.1.0
openai>=1.35.0

# --- Azure Services ---
azure-servicebus>=7.11.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0

# --- Configuration & Utilities ---
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
loguru>=0.7.0
httpx>=0.25.0
aiohttp>=3.9.0
tenacity>=8.2.0
orjson>=3.9.0

# --- Testing ---
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# --- Code Quality ---
ruff>=0.1.0
mypy>=1.7.0
